#159456364 Decoded user id should be passed down

+++ server/controllers/myDiaryController.js
@@ -2,27 +2,27 @@ import client from '../models/database';
 
 class myDiaryController {
   static createEntry(req, res) {
-    const entry = {
-      userId: req.body.userId,
-      entrytitle: req.body.entrytitle,
-      entrycontent: req.body.entrycontent,
-    };
+    const { userId, entrytitle, entrycontent } = req.body;
     const query = {
-      text: 'INSERT INTO entries(userId, entrytitle, entrycontent) VALUES($1, $2, $3)',
-      values: [entry.userId, entry.entrytitle, entry.entrycontent],
+      text: 'INSERT INTO entries(userId, entrytitle, entrycontent) VALUES($1, $2, $3) RETURNING entryid, userid, entrytitle ,entrycontent',
+      values: [userId, entrytitle, entrycontent],
     };
-    client.query(query).then(() => res.status(201).json({
-      success: true,
-      message: 'Entry Successfully created',
-      entry,
-    })).catch(error => res.status(500).json({ message: error.message }));
+    client.query(query)
+      .then((entry) => {
+        res.status(201).json({
+          success: true,
+          message: 'Entry Successfully created',
+          entry: entry.rows,
+        });
+      }).catch(error => res.status(500).json({ message: error.message }));
   }
 
   static getAllEntries(req, res) {
-    client.query('SELECT * FROM entries')
+    const { userId } = req.body;
+    client.query('SELECT * FROM entries where userid = $1', [userId])
       .then((entry) => {
-        if (entry.rowCount === 1) {
-          res.status(201).json({
+        if (entry.rowCount >= 1) {
+          res.status(200).json({
             success: true,
             message: 'entries successfully retrieved',
             entry: entry.rows,
@@ -30,7 +30,7 @@ class myDiaryController {
         }
         res.status(404).json({
           success: false,
-          message: 'No entry found',
+          message: 'Yet to create an entry',
         });
       })
       .catch(error => res.status(500).json({ message: error.message }));
@@ -57,10 +57,11 @@ class myDiaryController {
 
   static getEntryById(req, res) {
     const entryId = parseInt(req.params.entryId, 10);
-    client.query({ text: 'SELECT * FROM entries where entryId = ($1) ', values: [entryId] })
+    const { userId } = req.body;
+    client.query({ text: 'SELECT * FROM entries where entryId = $1 and userId = $2 ', values: [entryId, userId] })
       .then((entry) => {
         if (entry.rowCount === 1) {
-          res.status(201).json({
+          res.status(200).json({
             success: true,
             entry: entry.rows
           });
@@ -76,18 +77,15 @@ class myDiaryController {
 
   static editEntry(req, res) {
     const entryId = parseInt(req.params.entryId, 10);
-    const entry = {
-      entrytitle: req.body.entrytitle,
-      entrycontent: req.body.entrycontent,
-    };
+    const { userId, entrytitle, entrycontent } = req.body;
     const query = {
-      text: 'UPDATE entries SET entrytitle = ($1), entrycontent = ($2) WHERE entryId = ($3) RETURNING entrytitle ,entrycontent',
-      values: [entry.entrytitle, entry.entrycontent, entryId],
+      text: 'UPDATE entries SET entrytitle = ($1), entrycontent = ($2) WHERE entryId = ($3) and userId = ($4)  RETURNING entryid, entrytitle ,entrycontent',
+      values: [entrytitle, entrycontent, entryId, userId],
     };
     client.query(query)
       .then((editEntry) => {
         if (editEntry.rowCount === 1) {
-          res.status(201).json({
+          res.status(200).json({
             success: true,
             message: 'successfully edited',
             entry: editEntry.rows,
@@ -103,11 +101,12 @@ class myDiaryController {
   }
 
   static deleteEntry(req, res) {
+    const { userId } = req.body;
     const entryId = parseInt(req.params.entryId, 10);
-    client.query('DELETE FROM entries WHERE entryId = ($1) RETURNING *', [entryId])
+    client.query('DELETE FROM entries WHERE entryId = $1 and userId = $2 RETURNING *', [entryId, userId])
       .then((entry) => {
         if (entry.rowCount === 1) {
-          res.status(201).json({
+          res.status(200).json({
             success: true,
             message: 'successfully deleted',
             entry: entry.rows,
